backend/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
APScheduler==3.10.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.7.0
bs4==0.0.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.1.4
cryptography==41.0.7
dataclasses-json==0.6.2
debugpy==1.8.0
decorator==5.1.1
distro==1.8.0
emoji==2.9.0
executing==2.0.1
filetype==1.2.0
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.0
google==3.0.0
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-texttospeech==2.14.2
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.2
grpcio-status==1.59.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
ipykernel==6.26.0
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.5.0
jupyter_core==5.5.0
langchain==0.0.335
langdetect==1.0.9
langsmith==0.0.64
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
numpy==1.26.2
openai==1.2.4
outcome==1.3.0.post0
packaging==23.2
parso==0.8.3
pexpect==4.8.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.25.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.0
pydantic_core==2.14.1
pydub==0.25.1
Pygments==2.16.1
PyMySQL==1.1.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
rapidfuzz==3.5.2
regex==2023.10.3
requests==2.31.0
rsa==4.9
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.23
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.2.3
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
trio==0.23.1
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzlocal==5.2
unstructured==0.11.2
urllib3==2.0.7
wcwidth==0.2.9
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.2
